Add files via upload
This is a python script used to remove some of the tedium of starting a new Fortran program. When run, it prompts the user for some basic info on the program to be created, then creates a file with the appropriate name and extension. Currently, the only actual code created is the program beginning and end. The rest is all comments. There is a comment header containing some basic info about the program and some other comments to show where to put variables, subroutines, and the likes.